Scenario updates
- Refactor Scenario Struct for better readability.
- Identify bytes containing information for current
Chapter, Map, Auxiliary BGM, and Hidden Turn Limit.
- Add Enum listing byte addresses for each map.
- Identify bytes for changing the Label of the Victory
and Defeat Conditions(does not appear to have an effect
on gameplay).
- Add Enum listing byte addresses for each Victory and
Defeat condition.
- Reveal team modification-related byte. Functionality
requires more testing.
- Identify byte to set the level of any student you ask
for assistance or Recruit during a chapter. Special thanks
to Rengor1997 for pointing out the correlation.
- Identify Enemy Level Scaling bytes for each difficulty.
Special thanks to Raytwo for pointing me toward the first
one.
- Identify bytes controlling Expiration Dates for
Paralogues.
- Identify byte determining first available chapter for
Auxiliary Battle. Note that this includes the Brionac
Plateau shown to start at Chapter 2. This references the
Practice Battle tutorial auxiliary in addition to the
versions that begin appearing on Chapter 4.
- Identify bytes which reward Divine Pulses at the end of
a mission. Special thanks to Raytwo for finding a correlation
between Divine Pulses and these bytes.
- Reveal progress-related bytes. Functionality unknown.
- Identify three shorts controlling what units will be
unavailable in a scenario.
- Identify two shorts relating to unit IDs and their
Paralogues. Further research needed.
- Flesh out skeleton for second portion of fixed_scenario.

Signed-off-by: Kyle Huckins <kylehenglish@gmail.com>