Calendardata Information
- Create Struct "Celebration" to cover calendar events which contain
holiday, birthday, and an as-yet-unknown event, containing three
shorts calling for a character ID(used in Birthdays and type-3 events),
a byte to specify the month, day of the month, type of event, and
holiday activity(Hymn, Feast, or none), as well as one that determines
celebration visibility.
- Create Struct for Schedules, containing bytes determining if there's
been a rare monster sighting, which free day options are available,
what month and day it is, what day of the week it is, what food,
gardening, choir, or fishing event is abailable, and what tournaments
are available, and whether it is the first day of the month, a mission,
or a free day.
- Apply Celebration struct to first section of fixed_calendardata
- Apply Schedule struct to second, third, fourth, fifth, and sixth
  secions of fixed_calendardata.
- Identify five as-yet unknown sections of fixed_calendardata, and
  create skeleton structs for them.
- Add enums for Month, Celebration Activities, types of celebrations,
  food events, fishing events, gardening events, choir events, type of
  day, tournament types, and descriptions of known celebrations.

Signed-off-by: Kyle Huckins <kylehenglish@gmail.com>